Extract `mc-java-ksp` module
