+++ mc-java-ksp/build.gradle.kts
@@ -24,39 +24,27 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.tools.mc.java.routing.processor
+import io.spine.dependency.build.Ksp
+import io.spine.dependency.lib.Kotlin
+import io.spine.dependency.local.ProtoData
+import io.spine.dependency.local.TestLib
+import io.spine.dependency.local.ToolBase
 
-import com.google.devtools.ksp.symbol.KSFunctionDeclaration
-import com.google.devtools.ksp.symbol.Origin.JAVA
-import com.google.devtools.ksp.symbol.Origin.JAVA_LIB
+dependencies {
+    compileOnlyApi(Kotlin.Compiler.embeddable)
+    api(kotlin("stdlib"))
+    api(Ksp.symbolProcessingApi)
+    api(Ksp.gradlePlugin)
+    api(ProtoData.gradleApi)
 
-/**
- * Obtains the short name of the function.
- *
- * @returns just a name without braces.
- */
-internal val KSFunctionDeclaration.shortName: String
-    get() = simpleName.getShortName()
+    // The dependencies for the Gradle plugin part.
+    compileOnlyApi(gradleApi())
+    compileOnlyApi(gradleKotlinDsl())
+    compileOnlyApi(Kotlin.GradlePlugin.lib)
+    api(ToolBase.pluginBase)
 
-/**
- * Selects either diagnostic message depending on
- * the [origin][KSFunctionDeclaration.origin] of the declaration.
- *
- * For origins [JAVA] and [JAVA_LIB] the value of the [java] parameter is returned.
- * Otherwise, the [kotlin] string is returned.
- */
-internal fun KSFunctionDeclaration.msg(kotlin: String, java: String): String =
-    if (origin == JAVA || origin == JAVA_LIB) {
-        java
-    } else {
-        kotlin
-    }
-
-/**
- * Obtains the text for referencing this function in a diagnostic message.
- */
-internal val KSFunctionDeclaration.funRef: String
-    get() {
-        val shortRef = "`$shortName()`"
-        return msg("function $shortRef", "method $shortRef")
-    }
+    testImplementation(gradleTestKit())
+    testImplementation(gradleKotlinDsl())
+    testImplementation(Kotlin.GradlePlugin.lib)
+    testImplementation(TestLib.lib)
+}
