+++ mc-java-ksp/src/main/kotlin/io/spine/tools/mc/java/ksp/processor/KSFunctionDeclarationExts.kt
@@ -24,12 +24,39 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.tools.mc.java.routing.processor
+package io.spine.tools.mc.java.ksp.processor
 
-import com.google.devtools.ksp.symbol.KSDeclaration
+import com.google.devtools.ksp.symbol.KSFunctionDeclaration
+import com.google.devtools.ksp.symbol.Origin.JAVA
+import com.google.devtools.ksp.symbol.Origin.JAVA_LIB
 
 /**
- * Obtains the qualified name of this declaration or `null`
- * if the declaration does not have a qualified name.
+ * Obtains the short name of the function.
+ *
+ * @returns just a name without braces.
  */
-internal fun KSDeclaration.qualified(): String? = qualifiedName?.asString()
+public val KSFunctionDeclaration.shortName: String
+    get() = simpleName.getShortName()
+
+/**
+ * Selects either diagnostic message depending on
+ * the [origin][KSFunctionDeclaration.origin] of the declaration.
+ *
+ * For origins [JAVA] and [JAVA_LIB] the value of the [java] parameter is returned.
+ * Otherwise, the [kotlin] string is returned.
+ */
+public fun KSFunctionDeclaration.msg(kotlin: String, java: String): String =
+    if (origin == JAVA || origin == JAVA_LIB) {
+        java
+    } else {
+        kotlin
+    }
+
+/**
+ * Obtains the text for referencing this function in a diagnostic message.
+ */
+public val KSFunctionDeclaration.funRef: String
+    get() {
+        val shortRef = "`$shortName()`"
+        return msg("function $shortRef", "method $shortRef")
+    }
