+++ mc-java-routing/build.gradle.kts
@@ -31,9 +31,7 @@ import io.spine.dependency.lib.Kotlin
 import io.spine.dependency.lib.KotlinPoet
 import io.spine.dependency.local.CoreJava
 import io.spine.dependency.local.Logging
-import io.spine.dependency.local.ProtoData
 import io.spine.dependency.local.TestLib
-import io.spine.dependency.local.ToolBase
 import io.spine.dependency.test.Kotest
 import io.spine.dependency.test.KotlinCompileTesting
 
@@ -48,8 +46,6 @@ ksp {
 }
 
 dependencies {
-    compileOnly(Kotlin.Compiler.embeddable)
-
     // Dependencies for the code generation part.
     ksp(AutoServiceKsp.processor)?.because(
         "`RouteProcessorProvider` is annotated with `@AutoService`."
@@ -61,28 +57,15 @@ dependencies {
         the `AutoService` class on runtime.        
         """.trimIndent()
     )
-    implementation(kotlin("stdlib"))
-    implementation(Ksp.symbolProcessingApi)
-    implementation(Ksp.gradlePlugin)
+
     implementation(KotlinPoet.ksp)
-    implementation(ProtoData.gradleApi)
     implementation(CoreJava.server)
     implementation(project(":mc-java-base"))
+    implementation(project(":mc-java-ksp"))
 
-    // The dependencies for the Gradle plugin part.
-    compileOnly(gradleApi())
-    compileOnly(gradleKotlinDsl())
-//    compileOnly(Ksp.gradlePlugin)
-    compileOnly(Kotlin.GradlePlugin.lib)
-    implementation(ToolBase.pluginBase)
-
-    testImplementation(gradleKotlinDsl())
-    testImplementation(Kotlin.GradlePlugin.lib)
-//    testCompileOnly(Kotlin.Compiler.embeddable)
+    testImplementation(gradleTestKit())
     testImplementation(Kotest.assertions)
     testImplementation(KotlinCompileTesting.libKsp)
-    testImplementation(gradleTestKit())
-    testImplementation(TestLib.lib)
     testImplementation(Logging.testLib)
 }
