+++ mc-java-routing/src/main/kotlin/io/spine/tools/mc/java/routing/processor/CommandRouteSignature.kt
@@ -30,6 +30,8 @@ import com.google.devtools.ksp.symbol.KSFunctionDeclaration
 import com.google.devtools.ksp.symbol.KSType
 import io.spine.base.CommandMessage
 import io.spine.core.CommandContext
+import io.spine.tools.mc.java.ksp.processor.qualifiedRef
+import io.spine.tools.mc.java.ksp.processor.ref
 
 internal class CommandRouteSignature(
     environment: Environment
