+++ mc-java-routing/src/main/kotlin/io/spine/tools/mc/java/routing/processor/RouteFun.kt
@@ -31,6 +31,7 @@ import com.google.devtools.ksp.symbol.KSFunctionDeclaration
 import com.google.devtools.ksp.symbol.KSType
 import com.squareup.kotlinpoet.ClassName
 import com.squareup.kotlinpoet.ksp.toClassName
+import io.spine.tools.mc.java.ksp.processor.shortName
 
 /**
  * Provides information about a route function detected in the [declaringClass]
